refactor(desktop): house extracted updater modules in an updater/ directory
Move updater-config and updater-scheduler (sources and tests) into
src/main/updater/ and tests/main/updater/ so the split accumulates as a
directory instead of prefixed siblings. Pure renames plus import-path
updates; updater.ts keeps re-exporting the same surface.
