+++ apps/desktop/src/main/updater/config.ts
