+++ apps/desktop/tests/main/updater/config.test.ts
@@ -5,7 +5,7 @@ import { join } from "node:path";
 import { DESKTOP_UPDATE_CHANNELS, SIDECAR_SOURCES } from "@open-design/sidecar-proto";
 import { describe, expect, it } from "vitest";
 
-import { DESKTOP_UPDATE_ENV, resolveDesktopUpdaterConfig } from "../../src/main/updater-config.js";
+import { DESKTOP_UPDATE_ENV, resolveDesktopUpdaterConfig } from "../../../src/main/updater/config.js";
 
 function makeRoot(): string {
   return mkdtempSync(join(tmpdir(), "od-updater-config-test-"));
